+++ Fabric/src/main/java/com/yungnickyoung/minecraft/travelerstitles/config/ConfigDimensionsFabric.java
@@ -24,10 +24,10 @@ public class ConfigDimensionsFabric {
     public boolean renderShadow = true;
 
     @ConfigEntry.Gui.Tooltip(count = 3)
-    public float textYOffset = -32.0f;
+    public int textYOffset = -32;
 
     @ConfigEntry.Gui.Tooltip(count = 3)
-    public float textXOffset = 0;
+    public int textXOffset = 0;
 
     @ConfigEntry.Gui.Tooltip(count = 2)
     public String dimensionBlacklist = "[]";
