+++ Forge/src/main/java/com/yungnickyoung/minecraft/travelerstitles/config/ConfigBiomesForge.java
@@ -11,8 +11,8 @@ public class ConfigBiomesForge {
     public final ForgeConfigSpec.ConfigValue<String> textColor;
     public final ForgeConfigSpec.ConfigValue<Double> textSize;
     public final ForgeConfigSpec.ConfigValue<Boolean> renderShadow;
-    public final ForgeConfigSpec.ConfigValue<Double> textYOffset;
-    public final ForgeConfigSpec.ConfigValue<Double> textXOffset;
+    public final ForgeConfigSpec.ConfigValue<Integer> textYOffset;
+    public final ForgeConfigSpec.ConfigValue<Integer> textXOffset;
     public final ForgeConfigSpec.ConfigValue<String> biomeBlacklist;
     public final ForgeConfigSpec.ConfigValue<Integer> recentBiomeCacheSize;
     public final ForgeConfigSpec.ConfigValue<Boolean> centerText;
@@ -84,7 +84,7 @@ public class ConfigBiomesForge {
                 " If Horizontally Center Title is enabled, this number is relative to the center of the screen.\n" +
                 " If Horizontally Center Title is disabled, this number is relative to the top of the screen.\n" +
                 " Default: -33.0")
-            .define("Text Y Offset", -33.0);
+            .define("Text Y Offset", -33);
 
         textXOffset = BUILDER
             .comment(
@@ -92,7 +92,7 @@ public class ConfigBiomesForge {
                 " If Horizontally Center Title is enabled, this number is relative to the center of the screen.\n" +
                 " If Horizontally Center Title is disabled, this number is relative to the left side of the screen.\n" +
                 " Default: 0.0")
-            .define("Text X Offset", 0.0);
+            .define("Text X Offset", 0);
 
         biomeBlacklist = BUILDER
             .comment(
